us_penny_stock_scanner_mvp/requirements.txt
python-dotenv
requests
pandas
yfinance
rich

